Updated style sheet and updated paths
`common-definitions.R`, along with other tRNA project related stuff, is
going to be moved to the project directory. In anticipation of this,
paths to some data files have been adjusted.

The colors were updated to the agreed-on colour scheme, and the font was
set.

Fixed a bug regarding the consistent usage of stage names.